eb-final/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
boto3==1.34.113
botocore==1.34.113
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.6
faiss-cpu==1.8.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.4 
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-aws==0.1.6
langchain-community==0.2.1
langchain-core==0.2.1
langchain-google-genai==1.0.5
langchain-text-splitters==0.2.0
langsmith==0.1.63
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pypandoc==1.13
pandoc==2.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
rsa==4.9
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.30
tenacity==8.3.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.0


